pyproject.toml
[tool.poetry]
name = "tc-backend"
version = "0.1.0"
description = "Task-centric backend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
psycopg2-binary = "^2.9.7"
celery = "^5.3.4"
redis = "^5.0.1"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.7"

tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
